Move random data generation to core.util
